noteng_project/requirements.txt
Django==5.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
django-cloudinary-storage
drf-yasg
rest_framework_simplejwt
python-decouple